Issue #941. In order to export bamDepth (along with bamDepthMother, bamDepthFather), preprocessing code was added to CohortModel.promiseExportVariants. This code refreshes the variants to be exported with these field by finding the matching coverage region in the bam depth. This code that finds the matching coverage region was incorrectly looking for an exact match on the variant start position. The code was fixed to look for a matching region, as the bam depth records are binned to minimize traffic and memory consumption.
Code in model.VariantExporter had to be fixed to properly initialize the bamDepth fields for called variants. Exporting called variants in a vcf format is an edge case that requires special code that recalls the variants in order to capture the correct vcf header and records. This code also had to preserve the bamDepth fields that were obtained in the CohortModel preprocessing described above.

In summary, the bamDepth, bamDepthMother, and bamDepthFather fields are now exporting properly. Also, for clarity, SIFT and Polyphen columns have been removed from the exported dataset since they are no longer displayed in the variant detail panel.