Silent runtime error occurs when loading saved analysis file #967
This is caused by the JSON.stringify because flaggedVariants is an array with elements also in other DangerSummary.badges[filter] arrays. By cloning the variant before pushing it to flaggedVariants, we make these variants unique and ensure that they will not be bypassed (represented as an empty object in JSON.stringify) when the cached analysis file is created.