Genes with no transcripts #964
Show error message when user selects a gene with no transcripts.
Also, rework the alerts so that when the error is encountered, the alert panel is displayed and the alert is highlighted.

Also, fixed side nav bars height so that content is not truncated. (The vertical panels in the navigation drawers are cropped for approximately 10 pixels. #973)