Variant detail for a called variant shows incomplete info Pedigree genotype viz (in VariantInspectCard). #966
The problem occurred when the sampleName was blank in the SampleModel. This condition occurs when there is not a vcf file and thus, no selected sample. Instead we have an alignment file only. By looking at the bam header and grabbing the sample name, we can use this as a generated sample name to call variants and correctly parse the output.